v0.0.1 INITIAL VERSION
This is the initial commit for the ED solver library.
The actual code includes all the development of the
previous two versions: DMFT_ED, the generic ED code, and
DMFT_LANC_ED, the massively parallel ED code for the normal
phase only.

Everything has been imported here, with some simplfications.
The code has been re-structured and re-organized.

This is beta version. Still to be tested but it compiles.