v0.0.18 Small Fix in ED_MAIN
Fixed a small bug in the lattice (with basis) implementation of ED_INIT_SOLVER, for the use with REPLICA bath type.
Now, both the serial and MPI lattice implementations take the Hloc argument as optional, and the bath_type == "replica" case is handled properly (by calling the single-site implementation of init_solver and passing  to it the impHloc as preset from replica symmetry basis). 

For reference: the present fix enables correct handling of function overloading for ed_init_solver(comm,Bath) at line 136 of ed_kane_mele in the driver-database (commit id: 54d95ec). 

Testing ahead.