Added a small recursive subroutine called "shift_sector" inside ED_DIAG_NORMAL in order to keep track of a bigger number of sectors when ED_SECTORS=T and ED_SECTORS_SHIFT > 1. Previously, the code was able to change the indices of the sectors only one by one, so that starting from sector (3,3) with ED_SECTORS_SHIFT=2, it diagonalized only sectors like (4,3), (5,3) and (3,5). Now, starting from the sector (3,3) with ED_SECTORS_SHIFT=2, the code diagonalizes also sectors like (4,4).
This turns out to be useful when the true ground state of the system changes sector (with the
variation of the parameters) from (3,3) to (4,4) for example. Without this subroutine
the code had to make an intermediate step inside sector (4,3) before being able to reach
(4,4), but there is nothing ensuring that the lowest energy state of this intermediate sector
would be lower than the one of sector (3,3). So you could end up keeping a state as the ground
state of the system even if it is no more, leading to wrong hysteresis effects.