v0.1.0 >> ACTHUNG << CRITICAL UPDATE + Bug fix
This update fixes a possible critical bug affecting "replica" bath, typically in
the context of ed_mode='nonsu2".

The bug is the result of the commixture of Hloc, i.e. the local part of the
non-interacting Hamiltonian, with the Hreplica, i.e. the parametric expression of
each bath level Hamiltonian in the replica representation.

@LORENZO_CRIPPA: An additional, unexplained bug was found in fitgf_replica.
The mask for Hreplica was constructed using diagonal and uplo elements. Why was that?
I assume the matrix basis could in principle (though it should not probably) not include
diagonal elements.

>>SEE BELOW FOR SOLUTION AND INFO:

A bit of history to make things clear.
The 1st version of the replica bath used explicitly Hloc, so that to each non-zero
element of Hloc it corresponsed a parameter and a matrix having 1 in the corresponding
position and 0 elsewhere.
In the updated version of the replica bath we realized (how smart!) that specifying
a basis of matrices and parameters was very much more advantegeous to control the
solution, specially in presence of symmetry breaking.
In the updating process however we miss to clearly separate the concept of
Hreplica = sum_p lambda_p Gamma_p from that of Hloc==> impHloc used internally in the code.
This created a dangerous bug as in some lines Hloc was inherited from Hreplica:
impHloc = Hreplica_build(lambda)

This had two terrible consequences:
*) wiped out the correct impHloc overwriting it with a replica Hamiltonian
*) introduced a symmetry breaking amplitude explicitly in the impurity part, de facto
duplicating its effect (which we keep within the bath).

In this updated version we separated the two concepts. Which required a more
careful writing of the replica bath, Hreplica construction and fit. In addition
this required a bit of work to decide how to pass the information to setup Hreplica
AND Hloc.

>> SOLUTION AND IMPLEMENTATION
This is solved with a small breaking of back-compatibility in some ED routine.

*) ed_set_Hloc ceases to exist.
*) ed_set_Hreplica is used to setup the Hreplica and replaces ed_set_Hloc.
It includes wrapper for R-DMFT, inequivalent sites. This requires to define a number Nineq of
Hreplica_lambda array parameters in the shape [Nlat,Nbasis].
*) ed_get_bath_dimension takes as optional argument Hloc or Hreplica_basis (the matrices) is
bath_type="replica".
*) ed_init_solver DOES NOT take Hloc as input anymore. It only requires ([mpicomm],bath).
*) "replica" bath type is now fully supported for R-DMFT.
NOTE: All these steps are INDEPENDENT from Hloc setup.

*) ed_solve has a mandatory input Hloc. This is used to setup impHloc internally.
*) ed_solve for R-DMFT ineq sites has a new input mpi_lanc=T,F to select how to use MPI
environment: T=parallel lanczos, serial sites, F=serial lanczos, parallel sites.
*) ed_chi_fit_gf for R-DMFT cases has Hloc as mandatory input. This is required to setup impHloc
during the fit for different sites.