Intermediate commit + BUG FIX for build_op_Ns
* Developing evaluation of the impurity density matrix.
By now we only evaluate the diagonal part. The full calculation
requires a careful generation of the bath and impurity configurations
separately one from the other.

* BUG FIX: corrected a wrong do loop restricted to 1:Norb
rather than 1:Ns in build_op_Ns for the
ed_mode=superc,nonsu2. The bug went unnoticed because we use
this routine only to evaluate quantities for the impurity.