`Norb>1` test passed: comments
A single-point (fixed parameters, run until convergence with reasonable production input) test has been performed on the ED_BHZ_2D model.

- \rho_IMP is diagonal (by construction at the moment), so that its elements represent probabilities.

- Tr[\rho_IMP] =  1.d0

- The single probabilities have been cross checked with an alternative calculation (more info below), and we got a fair good correspondence of *all* values:

    This        Benchmark
    0.0012    0.0012
    0.0047    0.0047
    0.0047    0.0047
    0.0049    0.0049
    0.0049    0.0049
    0.0067    0.0067
    0.0071    0.0071
    0.0340    0.0340
    0.0340    0.0340
    0.0340    0.0340
    0.0340    0.0340
    0.0874    0.0873
    0.0874    0.0873
    0.0898    0.0899
    0.0898    0.0899
    0.4753    0.4754

- Some comments and prints have been added or modified in the code to aid the testing.

---------------
Additional Info
---------------

We (me and @frapao) have generated in an alternate way the diagonal elements by implementing the canonical density matrix definition directly on the pure-state basis for the impurity, and computing all the needed expectation values (similarly to how docc and nup, ndw are already computed in ED_OBSERVABLES). The code for that is not included here for it is very lengthy, ugly and ad-hoc, as it treats only the Norb=2 case, and works by construction for the diagonal case only.

To reproduce the test you can look at:

https://github.com/bellomia/FORK_DMFT_ED/commit/0915730f0e24e15113d9f6549b0e986d6f582eae

________

PS. The prior commit, claiming the test was failed, has been squashed in. The apparent missing probability was actually lost in the post-processing script. Sorry for that.