Start working on full algorithm
Some testing has been carried out on both HM_BETHE and BHZ_2D models. 

- For both the full algorithm gives a diagonal impDM, with some machine precision noise on a few off-diagonal terms (ranging from ~e-25 up to ~e-17), the others being a perfect 0.d0; for sure not a problem, rather a confirmation that we are really touching all the matrix elements and averaging to zero with excellent precision when Iimp/=Jimp.

- Scaling with Ns goes very badly, to have an idea:

       ~10 sec per loop with Ns=7 (sector dimension ~1.2k) [timed on BETHE, Nbath=6]
       ~15 min per loop with Ns=9 (sector dimension ~15k) [timed on BETHE, Nbath=8]
       ~1+ hour per loop with Ns=10 (sector dimension ~44k) [timed on BHZ, Nbath=4]

Optimization ahead: we'll try a slight modification of the Fock space map, to have just _one_ loop on the sector index {I}->{Iimp,Ibath}, and a nested loop just on just a impurity index{Jimp}, to be generated independently; a composition rule to retrieve the {J} index to select the correct state_dvector component from {Jimp,Ibath} has to be engineered. 
Currently instead we are decomposing the indices inside a global __double__ loop on the whole sector dimension, which is prohibitive for bitdimension{J} >> bitdimension{Jimp}. The optimization is possibile since we want to trace over the bath, so just one bath index is needed.