0.4.5 PATCH + BUG FIX
A small update plus fixing of a BUG in ED_OBSERVABLES_SUPERC

The mixed use of the procedure build_op_Ns created some misbehavior
in the calculation of the local energy terms, as some lines of code
disappeared (inside build_op_Ns) leaving unpredicatble behavior.
This was apparent in ED_OBSERVABLES_SUPERC where
* m = MAP(i) m in the Fock space
* ib = bdecomp(m,2Ns)
where both not present producing:
a) an error in application of the operator Cdg, C on the (missing)
state m
b) impossibility to check the Jcondition using Ib array.

Fixed here and checked in the other  ED_OBSERVABLES_{NORMAL,NONSU2}