0.4.6 BUG FIX in `ED_FIT_REPLICA`
We used to feed to the fit routine only the _uplo_ part of the mask of H_replica, which should make sense if hermiticity is guaranteed (and if we are sure the zeros in the self-energy matrix reflect the zeros in H_replica / in the GF).

Nevertheless:

- In ED_GF_NONSU2 we take a different strategy, namely to compute ALL components (without a mask of H_replica) if ed_all_g flag is set to T, else only the diagonal + the nonzero off-diagonal terms (via Hreplica_mask again)
- We cannot really be sure of the second assumption, namely that the self-energy follows the same symmetries of the GF.

So we changed the strategy in ED_FIT_REPLICA to match the one adopted in ED_GF_NONSU2 with the idea that:

- whenever we do not want to assume anything about the self-energies we set ed_all_g = T, and work on all the components of both the GFs and the SEs
- the strategy to save time if we want to exploit symmetries is the same on GFs and SEs, so to ensure a consistent treatment.

> This has fixed a *longstanding problem in the kane-mele model*, namely that ---in the full spin-SU(2) case [graphene]--- a "ed_all_g=T" calculation with a "eye+pauli_x bath" and a "eye+pauli_z bath" (with same spin-rotated initial guess) used to give a totally different AIM parameters after the first fit, despite having the same AIM groundstate after the first diagonalization (same eigensolutions).
> Now, with ed_all_g=T, we get the same AIM parameters and the same observables (even all the energies, kinetic and AIM included) to machine precision, after one loop.
> Results at self-consistency add a sensible additional layer of mystery and complexity (as usual) and are currently being investigated. A first look reveals a quite different convergence behaviour, yet a promising correspondence of observables and energies.

----------------------------------------

TODO(?)

Are we sure we want to fit all the components even when we know we can assume the symmetries of the GFs to apply to the SEs too?

> Even putting aside the raw computational cost, I'd argue that fitting more objects than necessary is a serious treat to the reliability of the fitting procedure itself, so maybe we should avoid whenever possibile to do that and exploit instead at least the hermiticity (i.e. the uplo option).

So, we could maybe separate the two assumptions and make a new flag, e.g. ed_all_self, to use *in both ED_FIT_REPLICA and ED_GF_NONSU2* like this:

if(ed_all_g)then
   Hmask=Hreplica_mask(wdiag=.false.,uplo=.true.)
   if(ed_all_self) Hmask=.true
else
   Hmask=Hreplica_mask(wdiag=.true.,uplo=.false.)
   if(ed_all_self) stop "ERROR: ed_all_self requires ed_all_g"
endif