0.4.7 BUGFIX
Modified the way the chemical potential is read for fixed density
calculations: previously, the presence of a "var.restart" file was
checked inside the "search_chemical_potential" function. This meant
that, if the user did not modify the potential by hand in the input
file, the first loop after a restart would be done with the original
value of XMU, potentially kicking the bath away and implying a big
number of loops just to get back to the previously obtained value.
Now the restart value of XMU is read after the input parameters in
ED_INPUT_VARS. A message is printed if the restart file is found. The
name of the restart file has changed to "xmu.restart" for clarity.

Bugfix in ED_MAIN: for the Real Space DMFT case, with MPI and
MPI_LANC=.false. the serial ed_solve is invoked. In that function,
"save_input_file" is called if MpiMaster=.true. However MpiMaster is
always true in this case, since the global variable is not touched. The
lattice solver function uses a local MPI_MASTER variable. Not knowing
exactly if this is just a legacy, the failsafe adopted solution is to
add an optional parameter "printflag" to ed_solve_single, which is true
by default. Its value is then set to MPI_MASTER when the function is
invoked in the lattice solver.
The trailing whitespace printing line in ed_solve_single is now also
subject to the conditional printflag.
This eliminates the recurring bug whereby, depending on the state of the
buffer, a lot of trailing spaces would be printed to the logfile at the
end of each loop and many cores would write to "used.inputED.in"
rendering it possibly unusable.