0.4.9 BUGFIX (dangerous)
Fixed a bug (with help from Adriano Amaricci) in the parallel version of the diagonalization
procedure, that triggered a stopping condition at the start of the Green's function calculation,
if some particular conditions were met (it was a tricky one...)

WARNING: the same bug can probably be found inside other versions of the code, like Cluster-DMFT

The bug concerns the diagonalization of sectors with both Dim > lanc_dim_threshold and DimDw < MpiSize.
In this case the code creates a group with only DimDw processes, while the remaining MpiSize-DimDw
should do nothing. The problem arises because the excluded processes are not aware of the actual
eigenvalues of the sector and they may fill their version of the state_list with zeros (since they
know nothing). So, long story short, the state_list after the diagonalization could refer to different
sectors inside different processes, leading to errors when the state_list is used (i.e. when the Green's
function calculation is performed).

The solution was very simple: adding a BCAST of the eigenvalues after the diagonalization of the
sector, so that also the excluded processes know those energy values and do not fill the state_list with
unwanted stuff.