BUGFIX:
This bug was found thanks to Alexander Kowalski, benchmarking EDIpack
with CTQMC. When ED_TOTAL_UD was not selected (orbital-resolved
calculation) in combination with the ED_TWIN symmetry, the check on
which sectors were to be discarded was incorrect: it was checked whether
any(Nups<Ndws): this caused, for example, a 2-orbital sector of the type
"2 3; 3 2" and its twin "3 2; 2 3" to be both excluded.

This fix proves successful in the considered case, 2 orbital Bethe
lattice. Since this situation may happen also in other versions of this
code (and especially EDIpack), I would suggest some testing before
applying it everywhere.

I only considered the "normal" version of the ED routine. SC and NONSU2
might need attention too.