0.5.1 PATCH
Improved replica bath init routine, based on reliable converged data and
comparison with the normal case: when the Hreplica_basis(isym) matrix is
diagonal we used to init with a linearly spaced offset, whose bandwidth
is controlled via input. Now we replace the lowest magnitude offset with
a smaller value (when applicable), so to enable an efficient coupling to
the low energy excitations in the impurity. This resembles much what is
currently done in the normal bath case, for the Ek (see lines 1118,1119)

Although I think this new "linspace,~zero,linspace" offset array would
work fine for every diagonal term in Hreplica, for now the change has
been applied just to the identity subcase, i.e. when the corresponding
lambda values correspond exactly to the Ek of the normal bath structure.
This should ensure full back compatibility.

Tested on the Kane Mele model, resulting in very clean runs.


-------

In general we might consider to give more freedom to the user, allowing
for a per-bath lambda initialization at the driver level, which amounts
to modify the code along the lines of what has been done in the CDMFT 
code: see https://github.com/QcmPlab/CDMFT-LANC-ED/tree/bath_overhaul