EXPERIMENT: move init_Hreplica_symmetries to assumed-rank lambda arrays
Huge simplification of set_Hreplica interface for symmetry expansion, by
means of assumed-rank arrays (Fortran 2018 feature).
> reference: §5.2 @ https://wg5-fortran.org/N1901-N1950/N1942.pdf

This allows to define a unique procedure to handle single-site and RDMFT
Hreplica initialization, with both the new and legacy APIs: the dispatch
to the specific implementation is done /at runtime/ via the select rank
construct, thus enabling the program to distinguish the [Nbath,Nsym] and
[Nineq,Nsym] cases in a concise, elegant way. Of course, the Nbath=Nineq
case is the tricky one and we handle it by prioritizing the new API.

Even though as a general principle static overloading may lead to faster
execution, here we call the procedure just once (at the solver init), so
I do not foresee significant performance problems.

————————————————————————————————————————————————————————————————————————

Tested with current (hash:7019911) database version of...

• ed_hm_square driver, which implements the new single-site API.

• ed_kane_mele_AFMxy driver, which implements the legacy RDMFT API.

In both cases everything seems to work fine.

> Fortran compiler used for testing: gcc v11.3.0 (Homebrew latest)

————————————————————————————————————————————————————————————————————————

Minimum version of gfortran should be 10.0 according to
> https://gcc.gnu.org/wiki/Fortran2018Status (see SELECT RANK)

For intel I could not find the exact minimum version, but for sure Intel
Fortran Compiler Classic 2021.1.1 supports all Fortran 2018 features.
> https://en.wikipedia.org/wiki/Intel_Fortran_Compiler

For reference, the latest Ulysses modules are respectively
- gcc version 11.2.1
- ifort version 2021.2.0
(but I have to investigate actual support for mkl and openmpi...)

I would say that merging this update may be reasonable, but we should
definitely have a discussion about it. So I push this commit to a new
branch: 'assumed_rank_lambdas'.