INTRODUCING BATH_TYPE=GENERAL
With this commit we introduce BATH_TYPE=GENERAL.
This is the most general bath and allows for a REPLICA-like
formulation of the bath hamiltonian with orbital- & spin-independent
hybridizations V_k. This bath_type contains all the feature of the replica one
and all ED_MODE=NORMAL/NONSU2/SUPERC are available.
At the present only ED_MODE=NORMAL has been tested.