1.0.4 minor update + bug fix + bug remains
Minor update and check.
Fixed a possible bug in ED_BATH_DIM.
BUG: to check Hermiticity of the bath Hamiltonian H=\sum_p l_p O_p
in the ED_BATH_DIM/get_bath_dimensions
the code implicitly calls the constructor Hreplica_build with
the given set of lambda (user input). However, in the case of RDMFT
the user specifies lambda_ineq and not lambda causing H=0 identically and
the check to fail.

SOLUTION: This is solved by allowing Hreplica_build to have optional argument
lambda. If given it uses that array to construct H. If not, it assumes lambda=1.
This is fine as far as the output H is used to check hermiticity or other symmetries only.

Remaining BUG:
apparently the bath dimension counter for the replica (/general) case using the
standard method (not based on symmetries) fails to return the correct number, which
is instead obtained using the symmetry based method.
> IF WE DECIDE THAT REPLICA/GENERAL REQUIRES THE USER TO SET THE HAMILTONIAN BASIS, SO
THAT NO IMPLICIT BASIS CAN BE CONSTRUCTED USING ANY GIVEN H_LOC TO BE DECOMPOED, THEN
THE ONLY MEANINGFUL WAY TO COUNT THE REQUIRED BATH DIMENSION IS THE SYMMETRY BASED ONE.
THE OTHER SHOULD RAISE AN ERROR OR FALLBACK TO THE SYMMETRY ONE.