Possible problem: Non-Physical Self-Energies on the real axis.
It may occur while solving multi-orbital problems with off-diagonal order that
the Self-Energy components on the real axis are non physical
(e.g. they acquire a positive imaginary part on the diagonal terms)

This problem is caused by the computation of the off-diagonal impurity Green's functions:
we compute Green's Functions using Lanczos method that gives access only to GFs like
	< \hat{O}^+ \hat{O} > = GF_O
therefore to compute off-diagonal terms we compute the GF for
	\hat{O} = c_iorb + c_jorb
( e.g. using subroutine lanc_build_gf_normal_mix )
to which we subtract the diagonal terms computed with another subroutine
( e.g. using subroutine lanc_build_gf_normal_diag)

In doing so we insert some numerical noise in the off-diagonal terms since the Krylov subspace
generated starting from \hat{O} |GS> is different from the ones generated starting from
c_iorb |GS> and c_jorb|GS>.

If one is looking at off-diagonal ordering, expecting a non-zero off-diagonal GF, the noise
is not a problem since it is dominated by the off-diagonal components but without off-diagonal
order (expecting a zero off-diagonal GF) the algorithm retrive small numbers that are pure noise
This noise then enters in the inversion of impG to determine impSigma and can lead to problems
like an unphysical Self-Energy.

This effect can be reduced by increasing the number of lanczos momenta LANC_NGFITER or smoothing
the peaks with a larger broadening EPS. Nonetheless some noise could remain when looking at states
without off-diagonal order but one could compute the self-energy from impG and impG0 files.

NOTE THAT this does not affect the self-energy on the Matsubara axis and the DMFT self-consistency
cycle since the noise is due to the multi-peak structure of the Green's Functions on the real axis
while on Matsubara, where the Green's functions are smooth, there are no problems