1.1.1 >ACTHUNG< BUG FIX BUT POSSIBLE ISSUES
This version is stable and contains the latest update.
Yet, the newly introduced use of assumed rank arrays, notably in
*ed_set_Hloc
*ed_chi2_fitgf

can possibly raise some misbehavior due to bad handling of the memory in
different compilers.
We are investigating this. Please report any inconsistency.

Example of misbehavior include:
. diagonalization returns bad values for some observables like the occupations
(this is caused by bad values of impHloc set by ed_set_Hloc)
. the fit of the weiss field returns NaN or unphysical values.
(this is caused by bad copy of the passed Weiss array).