1.3.1 BUG FIX
L.Crippa spotted a bug in the evaluation of the GF in the nonSU2 channel.
Following a wrong convention the GF is evaluated as <A+,A> rather than <A,A+>.
This led to a wrong algebraic manipulation of the off-diagonal components which
could potentially cause wrong calculations.

This was tested observing non-vanishing Sigma function component for U=0.
Some false negative have been observed, e.g. in the BHZ-Excitonic case where by
symmetry, the correct result was obtained anyhow.
A better check using SOC for t2g orbitals (from LC) showed incorrect results.

Patched changing sign in the algebraic manipulations of the off-digonal components.

Co-authored-by: lcrippa
Co-authored-by: gbeddalumia