BUGFIX:
case Real-space DMFT and mpi_lanc=F:

in ed_solve_single, check_MPI() is T but fmpi_=mpi_lanc=F. So the
MPI communicator is not set (as it should be). However the following
condition if(MpiMaster) is true for every core, and every core writes on
the used.input file, rendering it unreadable if not dangerous. Cases in
which binary blobs - probably coming from cores writing at the same
time- are present have been reported. If one uses this file as input for
a new run, the program either doesn't run or behaves erratically.

Fix for now:
if fmpi_ is F, don't print to used.input in ed_solve_single.
Only the master (not the same master of ed_solve_single's MpiComm) in
ed_solve_lattice will print, before ed_solve_single is called.

Fix for the future:
Enable in the lattice case the possibility of having different input
files for different atoms; this could expose the case where ULOC is
different per atom, which is now not accessible without specific coding
in the driver.