Experimental:
add a function to access variables that are assumed-shape arrays.
This works by setting up a fixed-shape "external" array which is
interoperable, and then every time the global variable is accessed by
the python code, update the "internal" fortran array to have the same
value.

it works, but it's also a complication and involves allocating two times
what is effectively the same quantity.
To be discussed.