Update:
added python interface for get_dens function. It works as a function, so
dens = ed.get_dens() with optional arguments ilat and iorb.

We add a global variable Nineq to the edi2py class, which is set upon
calling init_solver, so we know whether we are in real-space or
single-imp DMFT. This decides which get_dens function to call.

Template for all other get_observable functions.