Update:
now that we have the global variable Nineq, we can use to streamline the
functions that retrieve S and G_imp: they now work as python functions,
so the matrix is not passed as an argument. One writes e.g.

Smats = ed.get_sigma(ilat=#,axis=#,ishape=#)

where the arguments are optional and respectively:
-ilat, in the lattice case, if one wants a specific inequivalent site
-axis = "r" or "m" as before
-ishape: this is the new flag. For single-impurity, it's 3
[Nspin*Norb,Nspin*Norb,Lw] or 5 (default) [Nspin,Nspin,Norb,Norb,Lw].
For real-space dmft, it's 3 [Nineq*Nspin*Norb,Nineq*Nspin*norb,Lw], 4
[Nineq,Nspin*norb,Nspin*Norb,Lw] or 6 (default) [Nineq,Nspin,Nspin,Norb,Norb,Lw]