Update:
main: added finalize_solver function, it's the inverse of init_solver
setup: added delete_ed_structure, it's the inverse of init_ed_structure
edi2py: added interface to the two previous functions.

when finalize_solver is called, the effects of init_solver are reversed.
Not those of set_Hloc. This is not a problem, set_hloc can be called
again since as a first step it deallocates impHloc if this is allocated.