Intermediate commit:
this **effectively** fixes a bug which is not however understood.

The bug goes as follow:

if:
  using the python library
  ED_MODE is set to a string that contains the character "c"

the character is treated as EoF. Capital "C" has no problem, other
"character" variables have no problem.

examples:
"superc" -> "super"
"superC" -> "superC"
"superx" -> "superx"
"abcdef" -> "ab"
"cabdef" -> *nothing*

The bug manifests when the function parse_input_variable (from SciFortran)
is called for ED_MODE in ED_INPUT_VARS. It is an internal Fortran procedure,
but it only manifests itself when the library is used by a python script
through edipy2.

The origin of the bug is still obscure to me. This bruteforce patch puts
ED_MODE="superc" if it is read as "super". This effectively fixes the
problem, but is no solution.

We need to investigate this further.