Update documentation:
the functions hreplica_build and hgeneral_build were defined in
ed_bath_aux, but declared as public in ed_bath_replica. This is no
problem in the code since the latter imports the former, but it's a bit
messy. Also it gave problems to the documentation parser, because it
treated them as two distinct declarations of functions with the same
name. Moved the forward declarations in the BATH_AUX mofule