Update:
fix an annoying bug: mathjax (responsible for rendering the latex
formulas) is not included in the html of the pages unless the :math:
environment is used at least one in them. The sidebar doesn't count. So,
in pages with only plain text, the title of the "chisquare fit" section
in the sidebar was not rendered.
Fixed by telling conf.py to include always all assets.