1.6.5 Updated Code.
Testing to SUPERC+{normal,replic,general}.

All operative modes are functioning now.
It is worth observing (without too much suprise) that for a multi-orbital superconductor
bath_type=general returns a better approximation with respect to bath_type=replica, i.e.
closer to the bath_type=normal solution (which we take here as a reference).

Here are the results for a 2 orbital model, with Hubbard only attraction U=-1 and a small
crystal field M=0.05 (just to raise degeneracy).

These are the differences in the SC order parametes (note we are forbidding the natural inter-orbital
value which would be non-zero in replica/general case!).
../general_minimize/full_convergence]$ cat {normal,replica,.}/phisc_last.ed
    AA              AB             BA               BB
    0.271656271    -0.000000000    -0.000000000     0.271656346
    0.025616675     0.000000000    -0.000000000     0.025616460
    0.228558318    -0.000000000    -0.000000000     0.228558318

Here not the lack of independence in Ek and Dk values in the normal bath (related to crystal field).
This suggests that a possible better basis for the SC order parameter \lambda_2 and Matrix 2 can be provided.
../general_minimize/full_convergence]$ head normal/hamiltonian.restart;cat  hamiltonian.restart
            #Ek_l1_s1              Dk_l1_s1              Vk_l1_s1             #Ek_l2_s1              Dk_l2_s1              Vk_l2_s1
  -2.284917438957E-01   -3.122752285250E-01    2.174715116852E-01   -2.718211729602E-01    5.330634435381E-01    3.354672901921E-01
   0.000000000000E+00    1.956957687793E-01    2.291065679547E-01    0.000000000000E+00    1.956956515902E-01    2.291062289824E-01
   2.718205861804E-01    5.330652704765E-01    3.354672239833E-01    2.284937361597E-01   -3.122740064199E-01    2.174716189559E-01

  2
   2.860016446315E-01    2.590779762883E-01   -2.422876191952E-01    1.482672684959E-01
   8.554140321002E-02    8.554140586071E-02    0.000000000000E+00    8.850796110127E-02
   2.590779750290E-01    2.860016414397E-01    2.422876282459E-01    1.482672596236E-01

( 1.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)
( 0.00, 0.00)  ( 1.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)
( 0.00, 0.00)  ( 0.00, 0.00)  (-1.00,-0.00)  ( 0.00,-0.00)
( 0.00, 0.00)  ( 0.00, 0.00)  ( 0.00,-0.00)  (-1.00,-0.00)

( 0.00, 0.00)  ( 0.00, 0.00)  ( 1.00, 0.00)  ( 0.00, 0.00)
( 0.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)  ( 1.00, 0.00)
( 1.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)
( 0.00, 0.00)  ( 1.00, 0.00)  ( 0.00, 0.00)  ( 0.00, 0.00)

Tests results are available on request.