1.6.6 Updated Code: RDM implementation
Introduced an almost complete implementation of the impurity Reduced Density Matrix calculation.

* RDM is evaluated in all ed_mode=normal,superc,nonsu2 using Sparse_Maps to accelerate the algorithm.
* The RDM is accessed using ED_IO routines in ED_IO/get_rdm using the same notation as in C-DFMT code.
X This version is missing the implementation of further orbital reduction (orbital tracing) of the RDM
which is available only in the normal mode. Working on that, also using reading and reducing strategy.
Not sure if the RDM is actually saved to file. If not also this should be done to avoid losing work.