Intermediate commit:
case bath_type=normal: in nonsu2, this case was treated differently upon
building sigma. The only difference is only some elements of the G
matrix were assigned a value. Then, the inv() subroutine was called as
before.

There was an index bug somewhere in these for loops. Since however
Gmats, Greal and invG0 are all initialized to zero, and inv() is called
anyway, it seems to me it makes little sense to single out the bath normal case.

Removed the special case for normal. Now all the build_impg and impsigma
functions work.