1.8.0 Updated version, corrected RDM algorithm for superc/nonsu2 In this version we corrected the algorithm to evaluate the RDM in the superc/nonsu2 channels. In these cases there are fundamental fermionic signs to take into account related to the exchange of operators required to perform the bath trace.
The math explanation is rather convoluted and will be reported in a
specific paper asap.

The matter is that: RDM_superc/nonsu2 should have non-diagonal elements
which average to zero if the correct sign of the trace is not counted.
Those diagonal elements should be equal to (modulo a minus sign) the
order parameters, e.g. <c_up.c_dw> or <cdg_up.cdw>.