1.8.1 Bug fix. This version fixes a bug introduced in the last commit 1.8.0 with the correct RDM for superc/nonsu2 case.
The source of the bug was in the wrong counting of the bath states
ibath = ibathup + ibathdw*2**{Nb}
where Nb counts the number of bits corresponding to the bath in the Fock state.
The bug was to set Nb=Nbath (which counts the correct bits only for Norb=1).
The correct value is Nb=Norb*Nbath (the number of electronic levels of the bath).